command_center/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.0
Pillow>=10.0.0
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
minio>=7.1.0
bcrypt>=4.1.0